Add files via upload
It is always interesting to have a tool to track errors in sql, because :



An application error can happen and nothing is seen in MSTR, DBM1 ..
Optimize your code as unnecessary and repeated errors is waste of CPU  
It happens to us recently, a DBA launched a creation of triggers in Production as requested by the development, unfortunately the triggers don't work correctly driving the triggering SQL statements (inserts, updates) to have negative sqlcode. As the negative sqlcode is not well signaled in the program, the application support doesn't know what happens to the programs and  15 long minutes  happen between the alert and the problem resolution, and it was by chance because the DBA is here to check his triggers and realized that there is a problem.

We had Apptune , but it was disabled to save some CPU consumption. 

So i decided to write this Rexx to do exactly what Apptune does : Start a trace and then IFI Read from Opx buffers.


IFCID 053 allows to get the sqlcode of a SQL statement, among others things, like the stats related to the sql (number of getpages, wait times ...) - IFCID 058 is the same and has the same assembler macro, but IFCID58 must be started with IFCID numbers from 59 to 66 (IFCID 58 is an END SQL event, so you must have a corresponding Start SQL IFCID launched) . Also, and the most important thing, IFCID 53 reports the SQLCODE even if the Sql could not be executed (-551 or -805 for example) 

 

Use this program wisely, it does a loop on READA to scan every sqlcode on the system, so it consumes  CPU !! Be behind your screen and stop it quickly after you get the data you need.

Also, remember that all monitors must run in a priority higher than Db2, so if you can't afford this, be sure that you have enough priority to catch things, otherwise it is possible that you lose records ... In my shop, we have quite enough priority to get the data we need, otherwise i suggest that you create a STC with the right priority , or use one that exists by replacing the jcl ...